+++ library_manager_server/library_manager_server/Controllers/Loans.cs
@@ -27,7 +27,7 @@ public class Loans(ILibraryManager libraryManager, ILogger<Loans> logger, ISessi
         _logger.LogInformation("GetLoans called");
         string? sessionId = _sessionHandler.GetSession(HttpContext);
         if(sessionId == null) return Unauthorized();
-        double? userId = _sessionHandler.GetUserId(sessionId);
+        long? userId = _sessionHandler.GetUserId(sessionId);
         if(userId == null) return Unauthorized();
         return _libraryManager.GetLoans(userId.Value);
     }
@@ -35,7 +35,7 @@ public class Loans(ILibraryManager libraryManager, ILogger<Loans> logger, ISessi
     [HttpGet("{loanId:double}", Name = GetLoanName),]
     [ProducesResponseType(StatusCodes.Status200OK)]
     [ProducesResponseType(StatusCodes.Status401Unauthorized)]
-    public async Task<ActionResult<Model.Loan>> GetLoan(double loanId)
+    public async Task<ActionResult<Model.Loan>> GetLoan(long loanId)
     {
         _logger.LogInformation("GetLoan called");
         string? providedSessionId = _sessionHandler.GetSession(HttpContext);
@@ -44,7 +44,7 @@ public class Loans(ILibraryManager libraryManager, ILogger<Loans> logger, ISessi
             _logger.LogDebug("No session found");
             return Unauthorized();
         }
-        double? userIdOrnull = _sessionHandler.GetUserId(providedSessionId);
+        long? userIdOrnull = _sessionHandler.GetUserId(providedSessionId);
         if (userIdOrnull == null)
         {
             _logger.LogDebug("No user associated with this session");
@@ -70,9 +70,9 @@ public class Loans(ILibraryManager libraryManager, ILogger<Loans> logger, ISessi
         _logger.LogInformation("Creating loan {isbn}", isbn);
         string? sessionId = _sessionHandler.GetSession(HttpContext);
         if(sessionId == null) return Unauthorized();
-        double? userIdOrnull = _sessionHandler.GetUserId(sessionId);
+        long? userIdOrnull = _sessionHandler.GetUserId(sessionId);
         if (userIdOrnull == null){ return Unauthorized(); } 
-        double userid = userIdOrnull.Value;
+        long userid = userIdOrnull.Value;
         if (_libraryManager.HasActiveLoan(isbn))
         {
             _logger.LogInformation($"Loan {isbn} already active");
@@ -94,12 +94,12 @@ public class Loans(ILibraryManager libraryManager, ILogger<Loans> logger, ISessi
     [ProducesResponseType(StatusCodes.Status401Unauthorized)]
     [ProducesResponseType(StatusCodes.Status403Forbidden)]
     [ProducesResponseType(StatusCodes.Status400BadRequest)]
-    public async Task<IActionResult> DeleteLoans(double loanId)
+    public async Task<IActionResult> DeleteLoans(long loanId)
     {
         _logger.LogInformation("Deleting loan {loanId}", loanId);
         string? sessionId = _sessionHandler.GetSession(HttpContext);
         if(sessionId == null) return Unauthorized();
-        double? userId = _sessionHandler.GetUserId(sessionId);
+        long? userId = _sessionHandler.GetUserId(sessionId);
         if (userId == null){ return Unauthorized(); } 
         bool ret = _libraryManager.OwnsLoan(loanId, userId.Value);
         if(ret == false) return Forbid();
